Question_Generation_SIH/requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
langchain_community
fastapi
uvicorn